Add files via upload
A very basic layout done with only the form being made. Adding back-end elements to create a multi-functional message board and forum page for the students residing in the Oori hostel which encourages with meeting with other people having similar interests without the necessity of being chronically online or a social butterfly.